Update readme
Aspiring Data Analyst with a strong passion for working with data. Skilled in analyzing and visualizing data to find useful insights. Eager to learn and apply data analytics to solve real-world problems. Familiar with SQL, Excel, and data visualization tools. Looking for opportunities to grow and contribute in the field of data analytics